+++ sources/sphere_intersect.c
@@ -6,13 +6,13 @@
 /*   By: vbarsegh <vbarsegh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/23 16:46:50 by vbarsegh          #+#    #+#             */
-/*   Updated: 2024/11/23 16:47:36 by vbarsegh         ###   ########.fr       */
+/*   Updated: 2024/11/24 21:51:50 by vbarsegh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minirt.h"
 
-float	sphere_intersect(t_vector cam_center, t_vector ray, t_sphere *sphere)
+double	sphere_intersect(t_vector cam_center, t_vector ray, t_sphere *sphere)
 {
 	//<D,D>x^2 + (2<OC,D>)x + <OC,OC> - r^2 = 0
 	//O is camera center
